taskflow/requirements.txt
Django>=5.2.10
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
celery>=5.6.2
redis>=7.1.0
python-decouple>=3.8
django-cors-headers>=4.3.1
drf-spectacular>=0.27.2
pytest-django>=4.8.0
factory-boy>=3.3.1
django-filter>=24.2